fix: Modernize RCTBridgeModule import, for Swift-modules compatibility
This fixes several compilation errors with "Duplicate interface
definition for class […]" when a project uses Expo 44+, which has
Swift modules.

I don't expect this to be a breaking change. Reportedly, it's safe
to use

  #import <React/*.h>

with React Native 40+; see comment linked below. This project
already uses an import of that form. See the line

  #import <React/RCTUtils.h>

in the same file changed in this commit.

Details:
  https://github.com/expo/expo/issues/15622#issuecomment-997225774